# public/swagger.json
{"components":{"schemas":{"core.gh.models.Owner":{"properties":{"login":{"type":"string"}},"required":["login"]},"core.gh.models.Repository":{"properties":{"name":{"type":"string"},"owner":{"$ref":"#/components/schemas/core.gh.models.Owner"}},"required":["name","owner"]},"core.gh.models.response.Error":{"properties":{"message":{"type":"string"}},"required":["message"]},"core.gh.models.response.Errors":{"properties":{"errors":{"type":"array","items":{"$ref":"#/components/schemas/core.gh.models.response.Error"}}},"required":["errors"]},"core.gh.models.Contributor":{"properties":{"name":{"type":"string"},"contributions":{"type":"integer","format":"int32"}},"required":["name","contributions"]}}},"openapi":"3.0.0","paths":{"/org/{org}/repositories":{"get":{"operationId":"getGhRepositories","tags":["github repositories"],"summary":"Get Github repositories from org","description":"Get all Github repositories from organization given the organization name.","parameters":[{"schema":{"type":"string"},"in":"path","name":"org","description":"Organization name","type":"string","required":true}],"responses":{"200":{"description":"GitHub repositories for the given organization.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/core.gh.models.Repository"}}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/core.gh.models.response.Errors"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/core.gh.models.response.Errors"}}}}}}},"/org/{org}/contributors":{"get":{"operationId":"getGhContributors","tags":["github contributors"],"summary":"Get Github contributors from org","description":"Get all Github contributors from an organization given the organization name.","parameters":[{"schema":{"type":"string"},"in":"path","name":"org","description":"Organization name","type":"string","required":true}],"responses":{"200":{"description":"GitHub contributors for the given organization.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/core.gh.models.Contributor"}}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/core.gh.models.response.Errors"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/core.gh.models.response.Errors"}}}}}}}},"produces":["application/json"],"consumes":["application/json"],"info":{"description":"Service to get all repositories and contributors from a GitHub oranization","title":"GitHub Service","version":"1.0.0-6491717"},"tags":[{"name":"routes"}]}